# Improve IndexOfAnyAsciiSearcher ARM throughput