Also solves an edge-case bug for byte overloads on ARM64 where we'd get false positives (negatives for Except) when haystack values were exactly 128 above a value in the needle.
Closes #78718

Before
|             Method | Length |       Needle |      Mean |     Error |
|------------------- |------- |------------- |----------:|----------:|
| IndexOfAny128_Char | 100000 | AlphaNumeric |  18.27 us |  0.035 us |
| IndexOfAny128_Byte | 100000 | AlphaNumeric |  11.37 us |  0.032 us |

After
|             Method | Length |       Needle |      Mean |     Error |
|------------------- |------- |------------- |----------:|----------:|
| IndexOfAny128_Char | 100000 | AlphaNumeric | 16.363 us | 0.3711 us |
| IndexOfAny128_Byte | 100000 | AlphaNumeric |  9.621 us | 0.0884 us |